Simplify embeds example
